backend/pyproject.toml
[tool.poetry]
name = "cooltiger-backend"
version = "0.1.0"
description = "FastAPI backend for Project CoolTiger"
authors = ["CoolTiger Team <team@cooltiger.ai>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
google-cloud-firestore = "^2.16.1"
google-cloud-storage = "^2.18.0"
google-cloud-pubsub = "^2.22.0"
firebase-admin = "^6.6.0"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
